chore: Add Docker configuration and Flask server setup
