+++ README.md
@@ -1,1 +1,4 @@
-my first README
+# Docker起動手順
+docker-compose -f docker/docker-compose.yml up
+
+
